+++ CHANGELOG.md
@@ -6,5 +6,6 @@ Please consult the changelog to inform yourself about breaking changes and secur
 
 ## [v0.1.0](https://github.com/Materials-Data-Science-and-Informatics/dev-utils/tree/v0.1.0) <small>(2023-04-25)</small> { id="0.1.0" }
 
-* First release
-
+- Added `update-codemeta` pre-commit hook to work around current codemetapy issues, see:
+    - github:proycon/codemetapy#38
+    - github:proycon/codemetapy#39